+++ workers/unity/Assets/Scripts/Workers/UnityClientConnector.cs
@@ -1,26 +1,52 @@
-﻿using Improbable.Gdk.Core;
+﻿using System;
+using Improbable.Gdk.Core;
 using Improbable.Gdk.PlayerLifecycle;
 using Improbable.Worker.CInterop;
+using UnityEngine;
 
 namespace BlankProject
 {
-    public class UnityClientConnector : DefaultWorkerConnector
+    public class UnityClientConnector : WorkerConnector
     {
         public const string WorkerType = "UnityClient";
-        
+
         private async void Start()
         {
-            await Connect(WorkerType, new ForwardingDispatcher()).ConfigureAwait(false);
+            var connParams = CreateConnectionParameters(WorkerType);
+            connParams.Network.ConnectionType = NetworkConnectionType.Kcp;
+
+            var builder = new SpatialOSConnectionHandlerBuilder()
+                .SetConnectionParameters(connParams);
+
+            if (!Application.isEditor)
+            {
+                var initializer = new CommandLineConnectionFlowInitializer();
+                switch (initializer.GetConnectionService())
+                {
+                    case ConnectionService.Receptionist:
+                        builder.SetConnectionFlow(new ReceptionistFlow(CreateNewWorkerId(WorkerType), initializer));
+                        break;
+                    case ConnectionService.Locator:
+                        builder.SetConnectionFlow(new LocatorFlow(initializer));
+                        break;
+                    case ConnectionService.AlphaLocator:
+                        builder.SetConnectionFlow(new AlphaLocatorFlow(initializer));
+                        break;
+                    default:
+                        throw new ArgumentOutOfRangeException();
+                }
+            }
+            else
+            {
+                builder.SetConnectionFlow(new ReceptionistFlow(CreateNewWorkerId(WorkerType)));
+            }
+
+            await Connect(builder, new ForwardingDispatcher()).ConfigureAwait(false);
         }
 
         protected override void HandleWorkerConnectionEstablished()
         {
             PlayerLifecycleHelper.AddClientSystems(Worker.World);
-        }
-
-        protected override string SelectDeploymentName(DeploymentList deployments)
-        {
-            return deployments.Deployments[0].DeploymentName;
         }
     }
 }