+++ workers/unity/Assets/Scripts/Workers/UnityGameLogicConnector.cs
@@ -2,17 +2,40 @@
 using Improbable.Gdk.Core;
 using Improbable.Gdk.PlayerLifecycle;
 using Improbable.Gdk.TransformSynchronization;
+using Improbable.Worker.CInterop;
+using UnityEngine;
 
 namespace BlankProject
 {
-    public class UnityGameLogicConnector : DefaultWorkerConnector
+    public class UnityGameLogicConnector : WorkerConnector
     {
         public const string WorkerType = "UnityGameLogic";
-        
+
         private async void Start()
         {
             PlayerLifecycleConfig.CreatePlayerEntityTemplate = CreatePlayerEntityTemplate;
-            await Connect(WorkerType, new ForwardingDispatcher()).ConfigureAwait(false);
+
+            IConnectionFlow flow;
+            ConnectionParameters connectionParameters;
+
+            if (Application.isEditor)
+            {
+                flow = new ReceptionistFlow(CreateNewWorkerId(WorkerType));
+                connectionParameters = CreateConnectionParameters(WorkerType);
+            }
+            else
+            {
+                flow = new ReceptionistFlow(CreateNewWorkerId(WorkerType),
+                    new CommandLineConnectionFlowInitializer());
+                connectionParameters = CreateConnectionParameters(WorkerType,
+                    new CommandLineConnectionParameterInitializer());
+            }
+
+            var builder = new SpatialOSConnectionHandlerBuilder()
+                .SetConnectionFlow(flow)
+                .SetConnectionParameters(connectionParameters);
+
+            await Connect(builder, new ForwardingDispatcher()).ConfigureAwait(false);
         }
 
         protected override void HandleWorkerConnectionEstablished()